+++ android/src/main/kotlin/com/smileidentity/flutter/Mapper.kt
@@ -45,6 +45,7 @@ import FlutterUploadImageInfo
 import FlutterUploadRequest
 import FlutterValidDocument
 import FlutterValidDocumentsResponse
+import com.smileidentity.flutter.utils.getCurrentIsoTimestamp
 import com.smileidentity.models.ActionResult
 import com.smileidentity.models.Actions
 import com.smileidentity.models.Antifraud
@@ -117,35 +118,35 @@ fun convertNonNullMapToNullable(map: Map<String, String>): Map<String?, String?>
 
 fun FlutterJobType.toRequest() =
     when (this) {
-        FlutterJobType.ENHANCED_KYC -> JobType.EnhancedKyc
-        FlutterJobType.DOCUMENT_VERIFICATION -> JobType.DocumentVerification
-        FlutterJobType.BIOMETRIC_KYC -> JobType.BiometricKyc
-        FlutterJobType.ENHANCED_DOCUMENT_VERIFICATION -> JobType.EnhancedDocumentVerification
-        FlutterJobType.SMART_SELFIE_ENROLLMENT -> JobType.SmartSelfieEnrollment
-        FlutterJobType.SMART_SELFIE_AUTHENTICATION -> JobType.SmartSelfieAuthentication
+        FlutterJobType.ENHANCEDKYC -> JobType.EnhancedKyc
+        FlutterJobType.DOCUMENTVERIFICATION -> JobType.DocumentVerification
+        FlutterJobType.BIOMETRICKYC -> JobType.BiometricKyc
+        FlutterJobType.ENHANCEDDOCUMENTVERIFICATION -> JobType.EnhancedDocumentVerification
+        FlutterJobType.SMARTSELFIEENROLLMENT -> JobType.SmartSelfieEnrollment
+        FlutterJobType.SMARTSELFIEAUTHENTICATION -> JobType.SmartSelfieAuthentication
     }
 
 fun JobType.toResponse() =
     when (this) {
-        JobType.EnhancedKyc -> FlutterJobType.ENHANCED_KYC
-        JobType.DocumentVerification -> FlutterJobType.DOCUMENT_VERIFICATION
-        JobType.BiometricKyc -> FlutterJobType.BIOMETRIC_KYC
-        JobType.EnhancedDocumentVerification -> FlutterJobType.ENHANCED_DOCUMENT_VERIFICATION
-        JobType.SmartSelfieEnrollment -> FlutterJobType.SMART_SELFIE_ENROLLMENT
-        JobType.SmartSelfieAuthentication -> FlutterJobType.SMART_SELFIE_AUTHENTICATION
+        JobType.EnhancedKyc -> FlutterJobType.ENHANCEDKYC
+        JobType.DocumentVerification -> FlutterJobType.DOCUMENTVERIFICATION
+        JobType.BiometricKyc -> FlutterJobType.BIOMETRICKYC
+        JobType.EnhancedDocumentVerification -> FlutterJobType.ENHANCEDDOCUMENTVERIFICATION
+        JobType.SmartSelfieEnrollment -> FlutterJobType.SMARTSELFIEENROLLMENT
+        JobType.SmartSelfieAuthentication -> FlutterJobType.SMARTSELFIEAUTHENTICATION
         else -> TODO("Not yet implemented")
     }
 
 fun FlutterJobTypeV2.toRequest() =
     when (this) {
-        FlutterJobTypeV2.SMART_SELFIE_AUTHENTICATION -> JobTypeV2.SmartSelfieAuthentication
-        FlutterJobTypeV2.SMART_SELFIE_ENROLLMENT -> JobTypeV2.SmartSelfieEnrollment
+        FlutterJobTypeV2.SMARTSELFIEAUTHENTICATION -> JobTypeV2.SmartSelfieAuthentication
+        FlutterJobTypeV2.SMARTSELFIEENROLLMENT -> JobTypeV2.SmartSelfieEnrollment
     }
 
 fun JobTypeV2.toResponse() =
     when (this) {
-        JobTypeV2.SmartSelfieAuthentication -> FlutterJobTypeV2.SMART_SELFIE_AUTHENTICATION
-        JobTypeV2.SmartSelfieEnrollment -> FlutterJobTypeV2.SMART_SELFIE_ENROLLMENT
+        JobTypeV2.SmartSelfieAuthentication -> FlutterJobTypeV2.SMARTSELFIEAUTHENTICATION
+        JobTypeV2.SmartSelfieEnrollment -> FlutterJobTypeV2.SMARTSELFIEENROLLMENT
         else -> TODO("Not yet implemented")
     }
 
@@ -224,14 +225,14 @@ fun FlutterUploadImageInfo.toRequest() =
 
 fun FlutterImageType.toRequest() =
     when (this) {
-        FlutterImageType.SELFIE_JPG_FILE -> ImageType.SelfieJpgFile
-        FlutterImageType.ID_CARD_JPG_FILE -> ImageType.IdCardJpgFile
-        FlutterImageType.SELFIE_JPG_BASE64 -> ImageType.SelfieJpgBase64
-        FlutterImageType.ID_CARD_JPG_BASE64 -> ImageType.IdCardJpgBase64
-        FlutterImageType.LIVENESS_JPG_FILE -> ImageType.LivenessJpgFile
-        FlutterImageType.ID_CARD_REAR_JPG_FILE -> ImageType.IdCardRearJpgFile
-        FlutterImageType.LIVENESS_JPG_BASE64 -> ImageType.LivenessJpgBase64
-        FlutterImageType.ID_CARD_REAR_JPG_BASE64 -> ImageType.IdCardRearJpgBase64
+        FlutterImageType.SELFIEJPGFILE -> ImageType.SelfieJpgFile
+        FlutterImageType.IDCARDJPGFILE -> ImageType.IdCardJpgFile
+        FlutterImageType.SELFIEJPGBASE64 -> ImageType.SelfieJpgBase64
+        FlutterImageType.IDCARDJPGBASE64 -> ImageType.IdCardJpgBase64
+        FlutterImageType.LIVENESSJPGFILE -> ImageType.LivenessJpgFile
+        FlutterImageType.IDCARDREARJPGFILE -> ImageType.IdCardRearJpgFile
+        FlutterImageType.LIVENESSJPGBASE64 -> ImageType.LivenessJpgBase64
+        FlutterImageType.IDCARDREARJPGBASE64 -> ImageType.IdCardRearJpgBase64
     }
 
 fun FlutterIdInfo.toRequest() =
@@ -271,7 +272,13 @@ fun FlutterEnhancedKycRequest.toRequest() =
         sourceSdk = "android (flutter)",
         timestamp = timestamp,
         signature = signature,
-        consentInformation = consentInformation.toRequest(),
+        consentInformation =
+            consentInformation?.toRequest() ?: ConsentInformation(
+                consentGrantedDate = getCurrentIsoTimestamp(),
+                personalDetailsConsentGranted = false,
+                contactInfoConsentGranted = false,
+                documentInfoConsentGranted = false,
+            ),
     )
 
 fun EnhancedKycResponse.toResponse() =
@@ -322,21 +329,21 @@ fun ActionResult.toResponse() =
         ActionResult.Completed -> FlutterActionResult.COMPLETED
         ActionResult.Approved -> FlutterActionResult.APPROVED
         ActionResult.Verified -> FlutterActionResult.VERIFIED
-        ActionResult.ProvisionallyApproved -> FlutterActionResult.PROVISIONALLY_APPROVED
+        ActionResult.ProvisionallyApproved -> FlutterActionResult.PROVISIONALLYAPPROVED
         ActionResult.Returned -> FlutterActionResult.RETURNED
-        ActionResult.NotReturned -> FlutterActionResult.NOT_RETURNED
+        ActionResult.NotReturned -> FlutterActionResult.NOTRETURNED
         ActionResult.Failed -> FlutterActionResult.FAILED
         ActionResult.Rejected -> FlutterActionResult.REJECTED
-        ActionResult.UnderReview -> FlutterActionResult.UNDER_REVIEW
-        ActionResult.UnableToDetermine -> FlutterActionResult.UNABLE_TO_DETERMINE
-        ActionResult.NotApplicable -> FlutterActionResult.NOT_APPLICABLE
-        ActionResult.NotVerified -> FlutterActionResult.NOT_VERIFIED
-        ActionResult.NotDone -> FlutterActionResult.NOT_DONE
-        ActionResult.IssuerUnavailable -> FlutterActionResult.ISSUER_UNAVAILABLE
+        ActionResult.UnderReview -> FlutterActionResult.UNDERREVIEW
+        ActionResult.UnableToDetermine -> FlutterActionResult.UNABLETODETERMINE
+        ActionResult.NotApplicable -> FlutterActionResult.NOTAPPLICABLE
+        ActionResult.NotVerified -> FlutterActionResult.NOTVERIFIED
+        ActionResult.NotDone -> FlutterActionResult.NOTDONE
+        ActionResult.IssuerUnavailable -> FlutterActionResult.ISSUERUNAVAILABLE
         ActionResult.IdAuthorityPhotoNotAvailable ->
-            FlutterActionResult.ID_AUTHORITY_PHOTO_NOT_AVAILABLE
+            FlutterActionResult.IDAUTHORITYPHOTONOTAVAILABLE
 
-        ActionResult.SentToHumanReview -> FlutterActionResult.SENT_TO_HUMAN_REVIEW
+        ActionResult.SentToHumanReview -> FlutterActionResult.SENTTOHUMANREVIEW
         ActionResult.Unknown -> FlutterActionResult.UNKNOWN
     }