+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release Notes
+
+## 10.3.5
+
+* Added enhanced SmartSelfie™ capture Selfie capture screen component
+* Added `skipApiSubmission` to SmartSelfie™ capture which defaults to `false` and will allow Selfie
+  capture without submission to the api
+* Make consent information optional on Biometric KYC, Enhanced KYC and Enhanced Document
+  Verification
+* Bump android to 10.5.2 (https://github.com/smileidentity/android/releases/tag/v10.5.2)
+* Bump iOS to 10.4.2 (https://github.com/smileidentity/ios/releases/tag/v10.4.2)
+
 ## 10.3.4
+
 * Added enhanced SmartSelfie™ capture to docV, enhanced docV, and biometric kyc
 * Added consent information to BioMetric KYC and Enhanced Document Verification
 * Bump android to 10.5.0 (https://github.com/smileidentity/android/releases/tag/v10.5.0)
@@ -12,7 +24,7 @@
 ## 10.3.2
 
 * Fixed metadata version check to support versions below Kotlin 2.0
-* Fixed document capture view on android to properly handle showInstructions and showConfirmation 
+* Fixed document capture view on android to properly handle showInstructions and showConfirmation
   screens
 
 ## 10.3.1
@@ -23,7 +35,8 @@
 
 ## 10.3.0
 
-* Introduce screens for the new Enhanced Selfie Capture Enrollment and Authentication Products (Android).
+* Introduce screens for the new Enhanced Selfie Capture Enrollment and Authentication Products (
+  Android).
 * Migrate to Gradle Plugin setup
 * Update to Kotlin 2.0
 * Bump android to 10.4.0 (https://github.com/smileidentity/android/releases/tag/v10.4.0)